requirements.txt
APScheduler==3.6.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
decorator==5.1.1
idna==3.3
lxml==4.9.1
pathlib==1.0.1
pycparser==2.21
python-dotenv==0.20.0
python-telegram-bot==13.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
